# More efficient decoding of floats and doubles