TBD: add fast and medium paths for decoding of floats and doubles and then test with JDK 11 and JDK 17

Tested with Scala 3 and JDK-21 on Intel® Core™ i7-11800H:

#### Before
```txt
Benchmark                          (size)   Mode  Cnt        Score       Error  Units
ArrayOfBigDecimalsReading.zioJson     128  thrpt    5    60147.258 ±  1206.244  ops/s
ArrayOfBigIntsReading.zioJson         128  thrpt    5    44040.175 ±   423.306  ops/s
ArrayOfDoublesReading.zioJson         128  thrpt    5   149040.245 ±  1606.733  ops/s
ArrayOfFloatsReading.zioJson          128  thrpt    5   233975.363 ± 28846.808  ops/s
GeoJSONReading.zioJson                N/A  thrpt    5    10582.132 ±   139.006  ops/s
PrimitivesReading.zioJson             N/A  thrpt    5  3618949.962 ± 74998.630  ops/s
```

#### After
```txt
Benchmark                          (size)   Mode  Cnt        Score       Error  Units
ArrayOfBigDecimalsReading.zioJson     128  thrpt    5    63247.638 ±   891.067  ops/s
ArrayOfBigIntsReading.zioJson         128  thrpt    5    49032.193 ±   564.246  ops/s
ArrayOfDoublesReading.zioJson         128  thrpt    5   173701.182 ±  3114.107  ops/s
ArrayOfFloatsReading.zioJson          128  thrpt    5   332838.364 ±  8347.542  ops/s
GeoJSONReading.zioJson                N/A  thrpt    5    11618.211 ±   422.198  ops/s
PrimitivesReading.zioJson             N/A  thrpt    5  3970602.729 ± 32732.282  ops/s
```